back/requirements.txt
Django==4.2.3
djangorestframework==3.14.0
asgiref==3.7.2
django-cors-headers==4.2.0
django-environ==0.10.0
pytz==2023.3
sqlparse==0.4.4
django-csp==3.7
python-dotenv==1.0.0
Pillow==10.0.1
django-ckeditor==6.7.0
django-admin-sortable2==2.1.10